finish assignment 2.B
