+++ Chirp.CLI.Client/Chirp.sln
@@ -1,0 +1,48 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio Version 17
+VisualStudioVersion = 17.0.31903.59
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Chirp.CLI", "Chirp.CLI.csproj", "{90D2D2B7-A9ED-4DAD-AF40-39E3542872FE}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "SimpleDB", "SimpleDB\SimpleDB.csproj", "{F42652E5-6BA4-432D-A487-187BDBC92AF4}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Debug|x64 = Debug|x64
+		Debug|x86 = Debug|x86
+		Release|Any CPU = Release|Any CPU
+		Release|x64 = Release|x64
+		Release|x86 = Release|x86
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{90D2D2B7-A9ED-4DAD-AF40-39E3542872FE}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{90D2D2B7-A9ED-4DAD-AF40-39E3542872FE}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{90D2D2B7-A9ED-4DAD-AF40-39E3542872FE}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{90D2D2B7-A9ED-4DAD-AF40-39E3542872FE}.Debug|x64.Build.0 = Debug|Any CPU
+		{90D2D2B7-A9ED-4DAD-AF40-39E3542872FE}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{90D2D2B7-A9ED-4DAD-AF40-39E3542872FE}.Debug|x86.Build.0 = Debug|Any CPU
+		{90D2D2B7-A9ED-4DAD-AF40-39E3542872FE}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{90D2D2B7-A9ED-4DAD-AF40-39E3542872FE}.Release|Any CPU.Build.0 = Release|Any CPU
+		{90D2D2B7-A9ED-4DAD-AF40-39E3542872FE}.Release|x64.ActiveCfg = Release|Any CPU
+		{90D2D2B7-A9ED-4DAD-AF40-39E3542872FE}.Release|x64.Build.0 = Release|Any CPU
+		{90D2D2B7-A9ED-4DAD-AF40-39E3542872FE}.Release|x86.ActiveCfg = Release|Any CPU
+		{90D2D2B7-A9ED-4DAD-AF40-39E3542872FE}.Release|x86.Build.0 = Release|Any CPU
+		{F42652E5-6BA4-432D-A487-187BDBC92AF4}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{F42652E5-6BA4-432D-A487-187BDBC92AF4}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{F42652E5-6BA4-432D-A487-187BDBC92AF4}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{F42652E5-6BA4-432D-A487-187BDBC92AF4}.Debug|x64.Build.0 = Debug|Any CPU
+		{F42652E5-6BA4-432D-A487-187BDBC92AF4}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{F42652E5-6BA4-432D-A487-187BDBC92AF4}.Debug|x86.Build.0 = Debug|Any CPU
+		{F42652E5-6BA4-432D-A487-187BDBC92AF4}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{F42652E5-6BA4-432D-A487-187BDBC92AF4}.Release|Any CPU.Build.0 = Release|Any CPU
+		{F42652E5-6BA4-432D-A487-187BDBC92AF4}.Release|x64.ActiveCfg = Release|Any CPU
+		{F42652E5-6BA4-432D-A487-187BDBC92AF4}.Release|x64.Build.0 = Release|Any CPU
+		{F42652E5-6BA4-432D-A487-187BDBC92AF4}.Release|x86.ActiveCfg = Release|Any CPU
+		{F42652E5-6BA4-432D-A487-187BDBC92AF4}.Release|x86.Build.0 = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+EndGlobal
