--- Data/chirp_cli_db.csv
@@ -1,5 +1,0 @@
-Author,Message,Timestamp
-ropf,"Hello, BDSA students!",1690891760
-adho,"Welcome to the course!",1690978778
-adho,"I hope you had a good summer.",1690979858
-ropf,"Cheeping cheeps on Chirp :)",1690981487
