+++ SimpleDB/IDatabaseRepository.cs
@@ -1,0 +1,7 @@
+namespace SimpleDB;
+
+interface IDatabaseRepository<T>
+{
+    public IEnumerable<T> Read(int? limit = null);
+    public void Store(T record);
+}
